+++ sdk/formrecognizer/azure-ai-formrecognizer/samples/sample_train_model_without_labels.py
@@ -42,9 +42,6 @@ class TrainModelWithoutLabelsSample(object):
         # Default for begin_train_model is `use_training_labels=False`
         poller = form_training_client.begin_train_model(self.container_sas_url)
         model = poller.result()
-        if model.status == "invalid":
-            form_training_client.delete_model(model.model_id)
-            exit(1)
 
         # Custom model information
         print("Model ID: {}".format(model.model_id))