+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -7,6 +7,7 @@
 - `training_files` parameter of `begin_train_model` is renamed to `training_files_url`
 - `use_labels` parameter of `begin_train_model` is renamed to `use_training_labels`
 - `list_model_infos` method has been renamed to `list_custom_models`
+- A `HttpResponseError` is now raised if a model with `status=="invalid"` is returned from the `begin_train_model()` or `train_model()` methods
 
 **New features**